refactor: rename component: ButtonBar
