+++ src/app/components/ButtonBar.svelte
